Switch to GetFocusDelegate() in GetFocusableArea()
This was part of the DialogNewFocusBehavior feature, but it appears
to fix another delegatesFocus bug (crbug.com/324112201). This CL
enables this part of the behavior, which is to call
GetFocusDelegate() instead of FindFocusableElementInShadowHost(),
in Element::GetFocusableArea(). Those differ slightly in how they
handle `delegatesFocus`.

This change is gated behind a killswitch, and I added a test that
fails without this CL.

Fixed: 324112201
Change-Id: I097f40d0b3d4bbd0218d9fc87bfc00fec8397ccd