Add WPTs for auctionReportBuyersDebugModeConfig, fix crash identified
Adds Web Platform Tests for the feature that allows debug mode to be
specified for auctionReportBuyers reporting.

Also fixes a crash identified by these tests. When an
auctionReportBuyersDebugModeConfig is provided and enabled is false, but
a debugKey is still provided, a TypeError is correctly thrown. The
auction should also fail at that point, but before this change that did
not occur. Instead, a CHECK was later hit.

Bug: 41485586, 327247098
Change-Id: Ia843011ff1dd74a5e4ba48002f5bea1b77d074d0